pyproject.toml
[project]
name = "guardrailed-ai-prospector"
version = "0.1.0"
description = "An AI-powered Multi-Agent Sales Orchestrator using heterogeneous LLMs and Guardrails"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.29.0",
    "openai-agents>=0.13.0",
    "openai-guardrails>=0.2.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
